Revert "Add overlay to enable expanded desktop option for power menu"
This reverts commit 01bbea9480b5c86fddac3f901b21ad2552e8e6f0.

Breaks our compile, as with this commit it is in the overlay but not in the
base package (i.e. settings). This could be solved by forking the PA settings
but sinc Jesus wrote the expanded desktop in the first place I bet we don't
need this addition in the first place. I will experiment later.