Add Email.apk
The logcat shows tons of static when it cannot find an email app. I could leave
it to the users, but email is pretty basic. It is an AOSP package anyway. There
is likely a better way to do this, I am unsure why it is not included as is.

Change-Id: I9de167ee14b045820ed5aa7e0e188929e5546015